reorganize mobile and desktop panel
