+++ src/components/HURUmap/Panel/DesktopPanel/PanelItem.js
