+++ src/components/HURUmap/Panel/DesktopPanel/RichData.js
