+++ src/components/HURUmap/Panel/MobileTabPanel/PanelItem.js
